refactor(chat): split ChatPanel.tsx into focused sub-modules
Lift the eight nested components and module-scoped helpers out of the
3,329-line ChatPanel.tsx into siblings under src/ui/src/components/chat/,
shrinking ChatPanel.tsx to ~1,135 lines. memo() boundaries and the
ScrollContext provider/consumer wiring are preserved.

Closes #497
