+++ src/ui/src/components/chat/ChatPanel.tsx
@@ -1,19 +1,15 @@
-import React, { createContext, memo, useContext, useEffect, useRef, useState, useMemo, useCallback } from "react";
+import { useEffect, useMemo, useRef, useState, useCallback } from "react";
 import { useTranslation } from "react-i18next";
-import { HighlightedMessageMarkdown } from "./HighlightedMessageMarkdown";
-import { HighlightedPlainText } from "./HighlightedPlainText";
+import { GitBranch, LoaderCircle } from "lucide-react";
 import { ChatSearchBar } from "./ChatSearchBar";
-import { AlertCircle, FileText, GitBranch, LoaderCircle, Mic, Plus, RotateCcw, Send, Split, Square, X } from "lucide-react";
 import { useAppStore } from "../../stores/useAppStore";
-import type { ToolActivity, CompletedTurn } from "../../stores/useAppStore";
 import {
-  loadChatHistory,
   loadAttachmentData,
+  loadChatHistory,
   loadAttachmentsForSession,
-  readFileAsBase64,
   listCheckpoints,
-  loadCompletedTurns,
   listSlashCommands,
+  loadCompletedTurns,
   openReleaseNotes,
   openUsageSettings,
   recordSlashCommandUsage,
@@ -24,49 +20,23 @@ import {
   submitPlanApproval,
   getAppSetting,
   setAppSetting,
-  listWorkspaceFiles,
   clearConversation,
   readPlanFile,
   loadDiffFiles,
   forkWorkspaceAtCheckpoint,
 } from "../../services/tauri";
 import { applySelectedModel } from "./applySelectedModel";
-import { MODELS } from "./modelRegistry";
-import { roleClassKey, shouldRenderAsMarkdown } from "./messageRendering";
-import { StreamingContext } from "./StreamingContext";
 import { findLatestPlanFilePath } from "./planFilePath";
 import type { PermissionLevel } from "../../stores/useAppStore";
-import { open } from "@tauri-apps/plugin-dialog";
 import { reconstructCompletedTurns } from "../../utils/reconstructTurns";
 import { extractLatestCallUsage } from "../../utils/extractLatestCallUsage";
-import type { SlashCommand, FileEntry } from "../../services/tauri";
-import type { ChatMessage, ChatAttachment, AttachmentInput, PendingAttachment } from "../../types/chat";
-import { base64ToBytes } from "../../utils/base64";
-import {
-  SUPPORTED_IMAGE_TYPES,
-  SUPPORTED_DOCUMENT_TYPES,
-  SUPPORTED_ATTACHMENT_TYPES,
-  MAX_ATTACHMENTS,
-  maxSizeFor,
-  isTextFile,
-} from "../../utils/attachmentValidation";
-import { useTypewriter } from "../../hooks/useTypewriter";
-import { extractToolSummary, relativizePath } from "../../hooks/toolSummary";
-import { AgentQuestionCard } from "./AgentQuestionCard";
-import { PlanApprovalCard } from "./PlanApprovalCard";
-import { ComposerToolbar } from "./composer/ComposerToolbar";
-import { SegmentedMeter } from "./composer/SegmentedMeter";
-import { ContextPopover } from "./composer/ContextPopover";
-import { WorkspaceActions } from "./WorkspaceActions";
-import { SlashCommandPicker, filterSlashCommands } from "./SlashCommandPicker";
-import { PinnedCommandsBar } from "./PinnedCommandsBar";
-import { AttachMenu } from "./AttachMenu";
+import type { AttachmentInput, ChatMessage } from "../../types/chat";
+import { debugChat } from "../../utils/chatDebug";
 import {
   AttachmentContextMenu,
   buildAttachmentMenuLabels,
 } from "./AttachmentContextMenu";
 import { AttachmentLightbox } from "./AttachmentLightbox";
-import { MessageAttachment, isTextDataMediaType } from "./MessageAttachment";
 import {
   downloadAttachment,
   openAttachmentInBrowser,
@@ -76,180 +46,28 @@ import {
   isShareSupported,
   type DownloadableAttachment,
 } from "../../utils/attachmentDownload";
-import { FileMentionPicker, matchFiles } from "./FileMentionPicker";
 import {
-  describeSlashQuery,
   parseSlashInput,
   resolveNativeHandler,
 } from "./nativeSlashCommands";
-import { checkpointHasFileChanges, clearAllHasFileChanges, buildRollbackMap } from "../../utils/checkpointUtils";
-import {
-  assistantTextForTurn,
-  buildPlainTurnFooters,
-  findTriggeringUserIndex,
-} from "../../utils/chatTurnFooter";
-import { ThinkingBlock } from "./ThinkingBlock";
-import { CompactionDivider } from "./CompactionDivider";
-import { SyntheticContinuationMessage } from "./SyntheticContinuationMessage";
-import {
-  hasUltrathink,
-  renderUltrathinkText,
-  resolveUltrathinkEffort,
-} from "./ultrathink";
-import {
-  extractCompactionEvents,
-  parseCompactionSentinel,
-  parseSyntheticSummarySentinel,
-} from "../../utils/compactionSentinel";
+import { resolveUltrathinkEffort } from "./ultrathink";
+import { extractCompactionEvents } from "../../utils/compactionSentinel";
 import { PanelToggles } from "../shared/PanelToggles";
 import { SessionTabs } from "./SessionTabs";
-import { deriveTasks, processActivities, turnHasTaskActivity, hasTaskActivity } from "../../hooks/useTaskTracker";
-import type { TaskTrackerResult, TrackedTask } from "../../hooks/useTaskTracker";
+import { AgentQuestionCard } from "./AgentQuestionCard";
+import { PlanApprovalCard } from "./PlanApprovalCard";
+import { WorkspaceActions } from "./WorkspaceActions";
 import { ScrollToBottomPill } from "./ScrollToBottomPill";
 import { useStickyScroll } from "../../hooks/useStickyScroll";
-import { useVoiceInput } from "../../hooks/useVoiceInput";
-import { debugChat } from "../../utils/chatDebug";
-import {
-  insertTranscriptAtSelection,
-  shouldOpenVoiceSettingsForError,
-} from "../../utils/voice";
 import styles from "./ChatPanel.module.css";
-import caretStyles from "./caret.module.css";
-
-import { formatTokens } from "./formatTokens";
-
-function shouldDisable1mContext(modelId: string | null): boolean {
-  if (!modelId) return false;
-  const entry = MODELS.find((m) => m.id === modelId);
-  return entry ? entry.contextWindowTokens < 1_000_000 : false;
-}
-
-/** Format a duration in seconds as "15s" or "2m 34s". */
-function formatElapsedSeconds(secs: number): string {
-  if (secs < 60) return `${secs}s`;
-  const m = Math.floor(secs / 60);
-  const s = secs % 60;
-  return `${m}m ${s}s`;
-}
-
-/** Format a duration in milliseconds as "15s" or "2m 34s". Sub-second turns
- *  round up to "1s" so the footer always shows something meaningful. */
-function formatDurationMs(ms: number): string {
-  return formatElapsedSeconds(Math.max(1, Math.floor(ms / 1000)));
-}
-
-/**
- * Lazily renders a PDF first-page thumbnail.
- *
- * Accepts either `dataBase64` (optimistic/pre-loaded data) or `attachmentId`
- * (fetches the body from the backend on demand). Shows a loading pill with
- * the filename while the thumbnail generates.
- */
-function PdfThumbnail({ dataBase64, attachmentId, filename, className, onClick, onContextMenu }: {
-  dataBase64?: string;
-  attachmentId?: string;
-  filename: string;
-  className?: string;
-  /** Left-click handler. Used to open the PDF with the system's default
-   *  PDF viewer rather than the lightbox (which only renders images). */
-  onClick?: () => void;
-  /** Right-click handler. Wired so PDF thumbnails get the same Claudette
-   *  context menu (Download / Copy / Open) as image attachments rather than
-   *  WebKit's default image menu. See issue 430. */
-  onContextMenu?: (e: React.MouseEvent) => void;
-}) {
-  const [src, setSrc] = useState<string | null>(null);
-  useEffect(() => {
-    let cancelled = false;
-
-    (async () => {
-      let b64 = dataBase64;
-      // If no inline data, fetch on demand from the backend.
-      if (!b64 && attachmentId) {
-        b64 = await loadAttachmentData(attachmentId);
-      }
-      if (!b64 || cancelled) return;
-      const bytes = base64ToBytes(b64);
-      const { generatePdfThumbnail } = await import("../../utils/pdfThumbnail");
-      const url = await generatePdfThumbnail(bytes.buffer as ArrayBuffer, 300, attachmentId);
-      if (!cancelled) setSrc(url);
-    })().catch(() => {});
-
-    return () => { cancelled = true; };
-  }, [dataBase64, attachmentId]);
-
-  // Both the loading-state pill and the rendered first-page thumbnail
-  // need to be keyboard-actionable when an onClick is wired — without
-  // role/tabIndex/Enter+Space handling, non-mouse users can't open the
-  // PDF.
-  const interactiveProps = onClick
-    ? {
-        role: "button" as const,
-        tabIndex: 0,
-        onKeyDown: (e: React.KeyboardEvent) => {
-          if (e.key === "Enter" || e.key === " ") {
-            e.preventDefault();
-            onClick();
-          }
-        },
-        "aria-label": `Open ${filename}`,
-      }
-    : {};
-  if (!src) {
-    return (
-      <div
-        className={styles.messagePdf}
-        onClick={onClick}
-        onContextMenu={onContextMenu}
-        {...interactiveProps}
-      >
-        <FileText size={16} />
-        <span>{filename}</span>
-      </div>
-    );
-  }
-  return (
-    <img
-      src={src}
-      alt={filename}
-      className={className}
-      onClick={onClick}
-      onContextMenu={onContextMenu}
-      style={onClick ? { cursor: "zoom-in" } : undefined}
-      {...interactiveProps}
-    />
-  );
-}
-
-/** Semantic colors for tool names — makes tool activity scannable at a glance. */
-const TOOL_COLORS: Record<string, string> = {
-  Read: "var(--tool-read)",
-  Glob: "var(--tool-read)",
-  Grep: "var(--tool-read)",
-  Write: "var(--tool-write)",
-  Edit: "var(--tool-edit)",
-  Bash: "var(--tool-bash)",
-  WebSearch: "var(--tool-web)",
-  WebFetch: "var(--tool-web)",
-  Agent: "var(--tool-agent)",
-  AskUserQuestion: "var(--accent-primary)",
-};
-
-function toolColor(name: string): string {
-  return TOOL_COLORS[name] ?? "var(--text-muted)";
-}
-
-/** Context to pass sticky-scroll handler into streaming sub-components. */
-const ScrollContext = createContext<{
-  handleContentChanged: () => void;
-}>({ handleContentChanged: () => {} });
-
-// Stable empty arrays to avoid Zustand selector re-renders when data is undefined.
-// Without these, `?? []` / `|| []` creates a new reference on every store update,
-// causing Object.is to return false and triggering unnecessary component re-renders.
-const EMPTY_COMPLETED_TURNS: CompletedTurn[] = [];
-const EMPTY_ACTIVITIES: ToolActivity[] = [];
-const EMPTY_ATTACHMENTS: ChatAttachment[] = [];
+import { shouldDisable1mContext, formatElapsedSeconds } from "./chatHelpers";
+import { ScrollContext } from "./ScrollContext";
+import { StreamingThinkingBlock } from "./StreamingThinkingBlock";
+import { StreamingMessage } from "./StreamingMessage";
+import { MessagesWithTurns } from "./MessagesWithTurns";
+import { ToolActivitiesSection } from "./ToolActivitiesSection";
+import { CurrentTurnTaskProgress } from "./CurrentTurnTaskProgress";
+import { ChatInputArea } from "./ChatInputArea";
 
 export function ChatPanel() {
   const { t } = useTranslation("chat");
@@ -1316,2010 +1134,6 @@ export function ChatPanel() {
           onContextMenu={(e) => openAttachmentMenu(e, lightbox.attachment)}
         />
       )}
-    </div>
-  );
-}
-
-/**
- * Isolated thinking block — subscribes to streamingThinking to avoid
- * re-rendering ChatPanel on every thinking delta.
- */
-const StreamingThinkingBlock = memo(function StreamingThinkingBlock({
-  sessionId,
-  isStreaming,
-  searchQuery,
-}: {
-  sessionId: string;
-  isStreaming: boolean;
-  searchQuery: string;
-}) {
-  const thinking = useAppStore(
-    (s) => s.streamingThinking[sessionId] || ""
-  );
-  if (!thinking) return null;
-  return (
-    <ThinkingBlock
-      content={thinking}
-      isStreaming={isStreaming}
-      enableTypewriter
-      searchQuery={searchQuery}
-    />
-  );
-});
-
-/**
- * Isolated streaming message component — runs the typewriter reveal at a steady
- * rate while the agent streams, and keeps draining the latched text after
- * streamingContent clears so the transition to the completed message is smooth
- * (the just-added chat message is hidden behind pendingTypewriter until drain
- * completes).
- */
-const StreamingMessage = memo(function StreamingMessage({
-  sessionId,
-  isStreaming,
-  searchQuery,
-}: {
-  sessionId: string;
-  isStreaming: boolean;
-  searchQuery: string;
-}) {
-  const streaming = useAppStore(
-    (s) => s.streamingContent[sessionId] || ""
-  );
-  const pendingText = useAppStore(
-    (s) => s.pendingTypewriter[sessionId]?.text ?? ""
-  );
-  const finishTypewriterDrain = useAppStore((s) => s.finishTypewriterDrain);
-  const { handleContentChanged } = useContext(ScrollContext);
-
-  const fullText = streaming || pendingText;
-  const { displayed, showCaret } = useTypewriter(fullText, isStreaming);
-
-  useEffect(() => {
-    handleContentChanged();
-  }, [displayed, handleContentChanged]);
-
-  // Drain complete + we're in pending-typewriter phase → release the hidden
-  // completed message so it takes over visually without a jump. Also clears
-  // streamingThinking in the same store update so StreamingThinkingBlock
-  // unmounts atomically with the completed message unhiding.
-  useEffect(() => {
-    if (!showCaret && !streaming && pendingText) {
-      finishTypewriterDrain(sessionId);
-    }
-  }, [showCaret, streaming, pendingText, sessionId, finishTypewriterDrain]);
-
-  if (!displayed) return null;
-
-  return (
-    <div
-      className={`${styles.message} ${styles.role_Assistant}`}
-      aria-live="polite"
-      aria-busy={isStreaming}
-    >
-      <div className={styles.content}>
-        <StreamingContext.Provider value={isStreaming || pendingText.length > 0}>
-          <HighlightedMessageMarkdown content={displayed} query={searchQuery} />
-        </StreamingContext.Provider>
-        {showCaret && <span className={caretStyles.caret} aria-hidden="true" />}
-      </div>
-    </div>
-  );
-});
-
-/**
- * Render a single completed turn summary (collapsible tool call list).
- */
-function TurnSummary({
-  turn,
-  collapsed,
-  onToggle,
-  taskProgress,
-  assistantText,
-  onFork,
-  onRollback,
-  searchQuery,
-  worktreePath,
-}: {
-  turn: CompletedTurn;
-  collapsed: boolean;
-  onToggle: () => void;
-  taskProgress?: TaskTrackerResult;
-  /** Joined text from assistant messages in this turn, used by copy action.
-   *  When empty, the copy button is not rendered. */
-  assistantText: string;
-  /** Called when the user clicks fork. When undefined the fork button is not
-   *  rendered (e.g. remote workspaces, where the fork command cannot run). */
-  onFork?: () => void;
-  /** Called when the user clicks rollback. Undefined hides the button
-   *  (e.g. turn is running, or no checkpoint exists for this turn). */
-  onRollback?: () => void;
-  /** Active chat-search query. Force-expands this card when non-empty and
-   *  the query matches inside any of the contained activity summaries. */
-  searchQuery: string;
-  worktreePath?: string | null;
-}) {
-  const hasElapsed = typeof turn.durationMs === "number" && turn.durationMs > 0;
-  const hasTokens =
-    typeof turn.inputTokens === "number" && typeof turn.outputTokens === "number";
-  const hasCopy = assistantText.length > 0;
-  const hasFork = !!onFork;
-  const hasRollback = !!onRollback;
-  const showFooter = hasElapsed || hasTokens || hasCopy || hasFork || hasRollback;
-
-  // Force-expand if the query matches in any activity summary or the
-  // resolved tool-summary fallback. Without this, marks would land in
-  // detached DOM (the collapsed branch never renders), so the bar's
-  // counter would tick up but nothing visible would change.
-  // Match against the same relativized text we render — otherwise a query
-  // hitting only the stripped workspace prefix would force-expand with no
-  // visible highlight inside.
-  const queryHasMatch =
-    !!searchQuery &&
-    turn.activities.some((a) => {
-      const text = relativizePath(
-        a.summary || extractToolSummary(a.toolName, a.inputJson),
-        worktreePath,
-      );
-      return text.toLowerCase().includes(searchQuery.toLowerCase());
-    });
-  const isExpanded = !collapsed || queryHasMatch;
-
-  return (
-    <div className={styles.turnSummaryWrapper}>
-      <div
-        className={styles.turnSummary}
-        role="button"
-        tabIndex={0}
-        onClick={onToggle}
-        onKeyDown={(e) => {
-          if (e.key === "Enter" || e.key === " ") {
-            e.preventDefault();
-            onToggle();
-          }
-        }}
-      >
-        <div className={styles.turnHeader}>
-          <span className={styles.toolChevron}>
-            {isExpanded ? "⌄" : "›"}
-          </span>
-          <span className={styles.turnLabel}>
-            {turn.activities.length} tool call
-            {turn.activities.length !== 1 ? "s" : ""}
-            {turn.messageCount > 0 &&
-              `, ${turn.messageCount} message${turn.messageCount !== 1 ? "s" : ""}`}
-          </span>
-        </div>
-        {isExpanded && (
-          <div className={styles.turnActivities}>
-            {turn.activities.map((act: ToolActivity) => (
-              <div key={act.toolUseId} className={styles.toolActivity}>
-                <div className={styles.toolHeader}>
-                  <span className={styles.toolName} style={{ color: toolColor(act.toolName) }}>
-                    {act.toolName}
-                  </span>
-                  {(act.summary || act.inputJson) && (
-                    <span className={styles.toolSummary}>
-                      <HighlightedPlainText
-                        text={relativizePath(act.summary || extractToolSummary(act.toolName, act.inputJson), worktreePath)}
-                        query={searchQuery}
-                      />
-                    </span>
-                  )}
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-      {taskProgress && taskProgress.totalCount > 0 && (
-        <TaskProgressBar
-          completedCount={taskProgress.completedCount}
-          totalCount={taskProgress.totalCount}
-        />
-      )}
-      {showFooter && (
-        <TurnFooter
-          durationMs={turn.durationMs}
-          inputTokens={turn.inputTokens}
-          outputTokens={turn.outputTokens}
-          assistantText={hasCopy ? assistantText : undefined}
-          onFork={onFork}
-          onRollback={onRollback}
-        />
-      )}
-    </div>
-  );
-}
-
-/** Bottom-of-turn action row: elapsed time, copy output, fork, rollback.
- *  Rendered below the turn summary for every completed turn. */
-function TurnFooter({
-  durationMs,
-  inputTokens,
-  outputTokens,
-  assistantText,
-  onFork,
-  onRollback,
-  className,
-}: {
-  durationMs?: number;
-  inputTokens?: number;
-  outputTokens?: number;
-  assistantText?: string;
-  onFork?: () => void;
-  onRollback?: () => void;
-  className?: string;
-}) {
-  const { t } = useTranslation("chat");
-  const [copied, setCopied] = useState(false);
-  const copyTimeoutRef = useRef<number | null>(null);
-  useEffect(() => {
-    return () => {
-      if (copyTimeoutRef.current !== null) {
-        window.clearTimeout(copyTimeoutRef.current);
-      }
-    };
-  }, []);
-
-  const handleCopy = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (!assistantText) return;
-    navigator.clipboard
-      .writeText(assistantText)
-      .then(() => {
-        setCopied(true);
-        if (copyTimeoutRef.current !== null) {
-          window.clearTimeout(copyTimeoutRef.current);
-        }
-        copyTimeoutRef.current = window.setTimeout(() => setCopied(false), 1200);
-      })
-      .catch((err) => {
-        console.error("Copy to clipboard failed:", err);
-      });
-  };
-
-  const handleFork = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    onFork?.();
-  };
-
-  const handleRollback = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    onRollback?.();
-  };
-
-  const tokensNode =
-    typeof inputTokens === "number" && typeof outputTokens === "number" ? (
-      <span key="tokens" className={styles.turnFooterTokens}>
-        {formatTokens(inputTokens)} in · {formatTokens(outputTokens)} out
-      </span>
-    ) : null;
-
-  const elapsedNode =
-    typeof durationMs === "number" && durationMs > 0 ? (
-      <span key="elapsed" className={styles.turnFooterElapsed}>
-        {formatDurationMs(durationMs)}
-      </span>
-    ) : null;
-
-  const actionButtons: React.ReactNode[] = [];
-  if (assistantText) {
-    actionButtons.push(
-      <button
-        key="copy"
-        type="button"
-        className={styles.turnFooterButton}
-        onClick={handleCopy}
-        title={copied ? t("copy_output_done") : t("copy_output")}
-        aria-label={t("copy_agent_output_aria")}
-      >
-        {copied ? (
-          // Checkmark feedback for ~1.2s after successful copy.
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <polyline points="20 6 9 17 4 12"></polyline>
-          </svg>
-        ) : (
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <rect x="9" y="9" width="13" height="13" rx="2" ry="2"></rect>
-            <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"></path>
-          </svg>
-        )}
-      </button>,
-    );
-  }
-  if (onFork) {
-    actionButtons.push(
-      <button
-        key="fork"
-        type="button"
-        className={styles.turnFooterButton}
-        onClick={handleFork}
-        title={t("fork_workspace")}
-        aria-label={t("fork_workspace")}
-      >
-        <Split size={14} />
-      </button>,
-    );
-  }
-  if (onRollback) {
-    actionButtons.push(
-      <button
-        key="rollback"
-        type="button"
-        className={styles.turnFooterButton}
-        onClick={handleRollback}
-        title={t("rollback_turn")}
-        aria-label={t("rollback_turn")}
-      >
-        <RotateCcw size={14} />
-      </button>,
-    );
-  }
-
-  if (!tokensNode && !elapsedNode && actionButtons.length === 0) return null;
-
-  const hasMetadata = !!(tokensNode || elapsedNode);
-
-  return (
-    <div
-      className={`${styles.turnFooter}${className ? ` ${className}` : ""}`}
-      onClick={(e) => e.stopPropagation()}
-    >
-      {tokensNode}
-      {tokensNode && elapsedNode && (
-        <span className={styles.turnFooterDot} aria-hidden="true">·</span>
-      )}
-      {elapsedNode}
-      {hasMetadata && actionButtons.length > 0 && (
-        <span className={styles.turnFooterDot} aria-hidden="true">·</span>
-      )}
-      {actionButtons}
-    </div>
-  );
-}
-
-/** Inline progress bar rendered beneath a turn summary when tasks are present. */
-function TaskProgressBar({
-  completedCount,
-  totalCount,
-}: {
-  completedCount: number;
-  totalCount: number;
-}) {
-  const percent = totalCount > 0 ? Math.round((completedCount / totalCount) * 100) : 0;
-  const allDone = completedCount === totalCount;
-
-  return (
-    <div className={styles.taskProgressBar}>
-      <div className={styles.taskProgressTrack}>
-        <div
-          className={`${styles.taskProgressFill} ${allDone ? styles.taskProgressDone : ""}`}
-          style={{ width: `${percent}%` }}
-        />
-      </div>
-      <span className={styles.taskProgressLabel}>
-        {completedCount}/{totalCount} tasks
-      </span>
-    </div>
-  );
-}
-
-/**
- * Renders all messages interleaved with completed turn summaries at the correct
- * chronological position. Uses a single store subscription + useMemo to avoid
- * per-message selectors and redundant re-renders during streaming.
- */
-const EMPTY_CHECKPOINTS: import("../../types/checkpoint").ConversationCheckpoint[] = [];
-
-type RollbackModalData = {
-  workspaceId: string;
-  sessionId: string;
-  checkpointId: string | null;
-  messageId: string;
-  messagePreview: string;
-  messageContent: string;
-  hasFileChanges: boolean;
-};
-
-const MessagesWithTurns = memo(function MessagesWithTurns({
-  messages,
-  workspaceId,
-  sessionId,
-  isRunning,
-  onForkTurn,
-  onAttachmentContextMenu,
-  onAttachmentClick,
-  searchQuery,
-}: {
-  messages: ChatMessage[];
-  /** The enclosing workspace id — forwarded into rollback data so the modal
-   *  can target the correct workspace (distinct from the session id after
-   *  the multi-session refactor). */
-  workspaceId: string;
-  /** The active chat session id. All per-conversation store reads (turns,
-   *  checkpoints, attachments, etc.) are now keyed by session id. */
-  sessionId: string;
-  isRunning: boolean;
-  /** Handler invoked when the user forks a turn. Undefined disables the fork
-   *  button (e.g. for remote workspaces where the command cannot run). */
-  onForkTurn?: (checkpointId: string) => void;
-  /** Right-click handler on message-image attachments. Lifted to ChatPanel so
-   *  the context menu renders at the top of the component tree. The third
-   *  argument is the persisted attachment id, used to lazy-load bytes for
-   *  PDFs (whose data_base64 is stripped on hydration). */
-  onAttachmentContextMenu?: (
-    e: React.MouseEvent,
-    attachment: DownloadableAttachment,
-    attachmentId?: string,
-  ) => void;
-  /** Left-click handler on message-image attachments — opens the lightbox. */
-  onAttachmentClick?: (
-    e: React.MouseEvent,
-    attachment: DownloadableAttachment,
-  ) => void;
-  /** Active chat-search query (Cmd/Ctrl+F). Empty string when the bar is
-   *  closed; non-empty values trigger highlight wrappers on each message. */
-  searchQuery: string;
-}) {
-  const { t } = useTranslation("chat");
-  const completedTurns = useAppStore(
-    (s) => s.completedTurns[sessionId] ?? EMPTY_COMPLETED_TURNS
-  );
-  const toggleCompletedTurn = useAppStore((s) => s.toggleCompletedTurn);
-  const checkpoints = useAppStore(
-    (s) => s.checkpoints[sessionId] ?? EMPTY_CHECKPOINTS
-  );
-  const openModal = useAppStore((s) => s.openModal);
-  const showThinkingBlocks = useAppStore(
-    (s) => s.showThinkingBlocks[sessionId] === true
-  );
-  // While the typewriter is finishing the drain after streamingContent cleared,
-  // hide the just-added completed assistant message — StreamingMessage renders
-  // it in-place, so showing both would duplicate the text.
-  const pendingMessageId = useAppStore(
-    (s) => s.pendingTypewriter[sessionId]?.messageId ?? null
-  );
-  const chatAttachments = useAppStore(
-    (s) => s.chatAttachments[sessionId] ?? EMPTY_ATTACHMENTS
-  );
-  const worktreePath = useAppStore(
-    (s) => s.workspaces.find((w) => w.id === workspaceId)?.worktree_path
-  );
-
-  // Pre-build a Map keyed by message_id for O(1) lookup in the render loop.
-  //
-  // Agent-origin attachments are persisted with `message_id` set to the *user*
-  // message that triggered the turn (FK-cascade-safe). For display they
-  // belong with the *assistant* message of the same turn — i.e. the next
-  // Assistant message after the FK anchor in chronological order. This
-  // re-route happens here so the storage shape can stay simple.
-  const attachmentsByMessage = useMemo(() => {
-    // Single reverse pass: each User message maps to the most recent
-    // Assistant message that follows it. O(n) instead of O(n²).
-    const userToNextAssistant = new Map<string, string>();
-    let nextAssistantId: string | null = null;
-    for (let i = messages.length - 1; i >= 0; i--) {
-      const m = messages[i];
-      if (m.role === "Assistant") {
-        nextAssistantId = m.id;
-      } else if (m.role === "User" && nextAssistantId) {
-        userToNextAssistant.set(m.id, nextAssistantId);
-      }
-    }
-    const map = new Map<string, ChatAttachment[]>();
-    for (const att of chatAttachments) {
-      const targetId =
-        att.origin === "agent"
-          ? (userToNextAssistant.get(att.message_id) ?? att.message_id)
-          : att.message_id;
-      const list = map.get(targetId);
-      if (list) list.push(att);
-      else map.set(targetId, [att]);
-    }
-    return map;
-  }, [chatAttachments, messages]);
-
-  // Build an index: afterMessageIndex → array of (turn, globalIndex) pairs.
-  // Only recomputed when completedTurns changes, not on every streaming update.
-  const turnsByPosition = useMemo(() => {
-    const map: Record<number, Array<{ turn: CompletedTurn; globalIdx: number }>> = {};
-    completedTurns.forEach((turn, globalIdx) => {
-      const key = turn.afterMessageIndex;
-      (map[key] ??= []).push({ turn, globalIdx });
-    });
-    return map;
-  }, [completedTurns]);
-
-  const completedTurnPositions = useMemo(
-    () => new Set(completedTurns.map((turn) => turn.afterMessageIndex)),
-    [completedTurns],
-  );
-
-  const findTriggeringUserIdx = useCallback(
-    (afterMessageIndex: number) => {
-      return findTriggeringUserIndex(messages, afterMessageIndex);
-    },
-    [messages],
-  );
-
-  // Map user message index → checkpoint for rollback buttons.
-  // Each user message maps to the latest preceding checkpoint, with the first
-  // user message mapping to null so it can clear the whole conversation.
-  const rollbackCheckpointByIdx = useMemo(
-    () => buildRollbackMap(messages, checkpoints),
-    [messages, checkpoints],
-  );
-
-  const buildRollbackData = useCallback(
-    (userIdx: number): RollbackModalData | null => {
-      if (!rollbackCheckpointByIdx.has(userIdx)) return null;
-      const target = rollbackCheckpointByIdx.get(userIdx) ?? null;
-      const userMsg = messages[userIdx];
-      if (!userMsg) return null;
-      return {
-        workspaceId,
-        sessionId,
-        checkpointId: target ? target.id : null,
-        messageId: userMsg.id,
-        messagePreview: userMsg.content.slice(0, 100),
-        messageContent: userMsg.content,
-        hasFileChanges: target
-          ? checkpointHasFileChanges(target, checkpoints)
-          : clearAllHasFileChanges(checkpoints),
-      };
-    },
-    [checkpoints, messages, rollbackCheckpointByIdx, workspaceId, sessionId],
-  );
-
-  // Joined assistant text per turn, used by the "Copy output" action in the
-  // turn footer. CompletedTurn is only persisted for tool-using turns, so the
-  // slice starts at the nearest preceding user message instead of the previous
-  // CompletedTurn boundary.
-  const assistantTextByTurnId = useMemo(() => {
-    const map = new Map<string, string>();
-    for (const turn of completedTurns) {
-      const userIdx = findTriggeringUserIdx(turn.afterMessageIndex);
-      if (userIdx === -1) {
-        map.set(turn.id, "");
-        continue;
-      }
-      map.set(
-        turn.id,
-        assistantTextForTurn(messages, userIdx, turn.afterMessageIndex),
-      );
-    }
-    return map;
-  }, [completedTurns, findTriggeringUserIdx, messages]);
-
-  // Per-turn rollback data, keyed by turn.id. Completed turns are only
-  // persisted for tool-using turns, so the triggering user is the nearest
-  // user message before the completed turn boundary.
-  const rollbackByTurnId = useMemo(() => {
-    const result = new Map<string, RollbackModalData>();
-    for (const turn of completedTurns) {
-      const userIdx = findTriggeringUserIdx(turn.afterMessageIndex);
-      if (userIdx === -1) continue;
-      const data = buildRollbackData(userIdx);
-      if (data) result.set(turn.id, data);
-    }
-    return result;
-  }, [buildRollbackData, completedTurns, findTriggeringUserIdx]);
-
-  const buildOnRollback = (turnId: string) => {
-    if (isRunning) return undefined;
-    const data = rollbackByTurnId.get(turnId);
-    if (!data) return undefined;
-    return () => openModal("rollback", data);
-  };
-
-  const plainTurnFootersByPosition = useMemo(() => {
-    return buildPlainTurnFooters(
-      messages,
-      rollbackCheckpointByIdx,
-      completedTurnPositions,
-      checkpoints,
-    );
-  }, [checkpoints, completedTurnPositions, messages, rollbackCheckpointByIdx]);
-
-  const renderPlainTurnFooter = (position: number) => {
-    const data = plainTurnFootersByPosition.get(position);
-    if (!data) return null;
-    const rollbackData = buildRollbackData(data.userIdx);
-    const onRollback =
-      !isRunning && rollbackData
-        ? () => openModal("rollback", rollbackData)
-        : undefined;
-    const onFork =
-      data.forkCheckpointId && onForkTurn
-        ? () => onForkTurn(data.forkCheckpointId!)
-        : undefined;
-    const hasRenderableTokens =
-      typeof data.inputTokens === "number" &&
-      typeof data.outputTokens === "number";
-
-    if (
-      !data.assistantText &&
-      !data.durationMs &&
-      !hasRenderableTokens &&
-      !onFork &&
-      !onRollback
-    ) {
-      return null;
-    }
-
-    return (
-      <TurnFooter
-        key={`plain-turn-footer-${data.position}`}
-        durationMs={data.durationMs}
-        inputTokens={data.inputTokens}
-        outputTokens={data.outputTokens}
-        assistantText={data.assistantText}
-        onFork={onFork}
-        onRollback={onRollback}
-        className={styles.messageTurnFooter}
-      />
-    );
-  };
-
-  // Compute cumulative task progress at each turn index in a single O(n) pass.
-  // Carries taskMap/todoMap forward across iterations instead of re-slicing.
-  const taskProgressByTurn = useMemo(() => {
-    const map = new Map<number, TaskTrackerResult>();
-    const taskMap = new Map<string, TrackedTask>();
-    const todoMap = new Map<string, TrackedTask>();
-    const nextSyntheticId = { value: 1 };
-    let anyTasksSoFar = false;
-
-    for (let i = 0; i < completedTurns.length; i++) {
-      processActivities(completedTurns[i].activities, taskMap, todoMap, nextSyntheticId);
-      if (turnHasTaskActivity(completedTurns[i])) {
-        anyTasksSoFar = true;
-      }
-      if (anyTasksSoFar) {
-        const tasks = [...taskMap.values(), ...todoMap.values()];
-        const completedCount = tasks.filter((t) => t.status === "completed").length;
-        map.set(i, { tasks, completedCount, totalCount: tasks.length });
-      }
-    }
-    return map;
-  }, [completedTurns]);
-
-  useEffect(() => {
-    debugChat("MessagesWithTurns", "layout", {
-      workspaceId,
-      sessionId,
-      messageIds: messages.map((msg) => msg.id),
-      turnLayout: completedTurns.map((turn) => ({
-        id: turn.id,
-        afterMessageIndex: turn.afterMessageIndex,
-        postLastMessage: turn.afterMessageIndex >= messages.length,
-        toolCount: turn.activities.length,
-      })),
-    });
-  }, [workspaceId, sessionId, messages, completedTurns]);
-
-  const renderTurns = (position: number) => {
-    const entries = turnsByPosition[position];
-    if (!entries) return null;
-    return entries.map(({ turn, globalIdx }) => (
-      <TurnSummary
-        key={turn.id}
-        turn={turn}
-        collapsed={turn.collapsed}
-        onToggle={() => toggleCompletedTurn(sessionId, globalIdx)}
-        taskProgress={taskProgressByTurn.get(globalIdx)}
-        assistantText={assistantTextByTurnId.get(turn.id) ?? ""}
-        onFork={onForkTurn ? () => onForkTurn(turn.id) : undefined}
-        onRollback={buildOnRollback(turn.id)}
-        searchQuery={searchQuery}
-        worktreePath={worktreePath}
-      />
-    ));
-  };
-
-  return (
-    <>
-      {messages.map((msg, idx) => {
-        // Sentinel dispatch for System messages — must precede the generic
-        // message bubble so compaction/synthetic-summary messages render
-        // as their own dedicated components.
-        if (msg.role === "System" && msg.id !== pendingMessageId) {
-          const compaction = parseCompactionSentinel(msg.content);
-          if (compaction) {
-            return (
-              <React.Fragment key={msg.id}>
-                {renderTurns(idx)}
-                <CompactionDivider
-                  event={{
-                    ...compaction,
-                    timestamp: msg.created_at,
-                    afterMessageIndex: idx,
-                  }}
-                />
-              </React.Fragment>
-            );
-          }
-          const syntheticBody = parseSyntheticSummarySentinel(msg.content);
-          if (syntheticBody !== null) {
-            return (
-              <React.Fragment key={msg.id}>
-                {renderTurns(idx)}
-                <SyntheticContinuationMessage
-                  body={syntheticBody}
-                />
-              </React.Fragment>
-            );
-          }
-        }
-        // Default rendering for User, Assistant, and non-sentinel System messages.
-        return (
-        <React.Fragment key={msg.id}>
-          {renderTurns(idx)}
-          {msg.id === pendingMessageId ? null : (
-          <div className={`${styles.message} ${styles[roleClassKey(msg.role, msg.content)]}`}>
-            {msg.role === "User" && (
-              <div className={styles.roleLabel}>{t("you_label")}</div>
-            )}
-            {msg.role === "Assistant" && msg.thinking && showThinkingBlocks && (
-              <ThinkingBlock content={msg.thinking} isStreaming={false} searchQuery={searchQuery} />
-            )}
-            <div className={styles.content}>
-              {attachmentsByMessage.has(msg.id) && (
-                <div className={styles.messageImages}>
-                  {attachmentsByMessage.get(msg.id)!.map((att) => {
-                    if (att.media_type === "application/pdf") {
-                      return (
-                        <PdfThumbnail
-                          key={att.id}
-                          dataBase64={att.data_base64 || undefined}
-                          attachmentId={att.id}
-                          filename={att.filename}
-                          className={styles.messageImage}
-                          onClick={() => {
-                            (async () => {
-                              // Persisted attachments strip data_base64 on first
-                              // load to avoid IPC bloat — fetch on demand.
-                              let b64 = att.data_base64;
-                              if (!b64) {
-                                b64 = await loadAttachmentData(att.id);
-                              }
-                              await openAttachmentWithDefaultApp({
-                                filename: att.filename,
-                                media_type: att.media_type,
-                                data_base64: b64,
-                              });
-                            })().catch((err) =>
-                              console.error("Failed to open PDF:", err),
-                            );
-                          }}
-                          onContextMenu={(e) =>
-                            onAttachmentContextMenu?.(
-                              e,
-                              {
-                                filename: att.filename,
-                                media_type: att.media_type,
-                                data_base64: att.data_base64,
-                              },
-                              att.id,
-                            )
-                          }
-                        />
-                      );
-                    }
-                    if (isTextDataMediaType(att.media_type)) {
-                      return (
-                        <MessageAttachment
-                          key={att.id}
-                          attachment={att}
-                          handlers={{ onContextMenu: onAttachmentContextMenu }}
-                        />
-                      );
-                    }
-                    return (
-                      <img
-                        key={att.id}
-                        src={`data:${att.media_type};base64,${att.data_base64}`}
-                        alt={att.filename}
-                        className={styles.messageImage}
-                        onClick={(e) =>
-                          onAttachmentClick?.(e, {
-                            filename: att.filename,
-                            media_type: att.media_type,
-                            data_base64: att.data_base64,
-                          })
-                        }
-                        onContextMenu={(e) =>
-                          onAttachmentContextMenu?.(
-                            e,
-                            {
-                              filename: att.filename,
-                              media_type: att.media_type,
-                              data_base64: att.data_base64,
-                            },
-                            att.id,
-                          )
-                        }
-                      />
-                    );
-                  })}
-                </div>
-              )}
-              {shouldRenderAsMarkdown(msg.role) ? (
-                // Assistant + System: run through Markdown so plan-mode dumps,
-                // setup-script output, and other multi-line system notes
-                // preserve headings, lists, and code blocks instead of
-                // collapsing newlines into a single paragraph.
-                <HighlightedMessageMarkdown content={msg.content} query={searchQuery} />
-              ) : searchQuery ? (
-                // While the search bar is open, render user messages as plain
-                // highlighted text so matches inside them get marked. The
-                // ultrathink rainbow animation is suppressed in this mode —
-                // searchability wins over the easter egg.
-                <HighlightedPlainText text={msg.content} query={searchQuery} />
-              ) : (
-                renderUltrathinkText(msg.content, {
-                  animated: false,
-                  styles: {
-                    ultrathinkChar: styles.ultrathinkChar,
-                    ultrathinkCharAnimated: styles.ultrathinkCharAnimated,
-                  },
-                })
-              )}
-            </div>
-          </div>
-          )}
-          {renderPlainTurnFooter(idx + 1)}
-        </React.Fragment>
-        );
-      })}
-      {/* Turns that finalized after or at the last message index */}
-      {completedTurns
-        .map((turn, globalIdx) => ({ turn, globalIdx }))
-        .filter(({ turn }) => turn.afterMessageIndex >= messages.length)
-        .map(({ turn, globalIdx }) => (
-          <TurnSummary
-            key={turn.id}
-            turn={turn}
-            collapsed={turn.collapsed}
-            onToggle={() => toggleCompletedTurn(sessionId, globalIdx)}
-            taskProgress={taskProgressByTurn.get(globalIdx)}
-            assistantText={assistantTextByTurnId.get(turn.id) ?? ""}
-            onFork={onForkTurn ? () => onForkTurn(turn.id) : undefined}
-            onRollback={buildOnRollback(turn.id)}
-            searchQuery={searchQuery}
-            worktreePath={worktreePath}
-          />
-        ))}
-    </>
-  );
-});
-
-/**
- * Current tool activities section — subscribes to toolActivities for this workspace.
- * Isolated so streaming text changes don't cause re-renders here.
- */
-const ToolActivitiesSection = memo(function ToolActivitiesSection({
-  sessionId,
-  isRunning,
-  searchQuery,
-  worktreePath,
-}: {
-  sessionId: string;
-  isRunning: boolean;
-  searchQuery: string;
-  worktreePath?: string | null;
-}) {
-  const activities = useAppStore(
-    (s) => s.toolActivities[sessionId] ?? EMPTY_ACTIVITIES
-  );
-  const [collapsed, setCollapsed] = useState(true);
-
-  // Auto-collapse when a new turn starts (activities goes from 0 to non-zero)
-  const prevLengthRef = useRef(0);
-  useEffect(() => {
-    if (isRunning && activities.length > 0 && prevLengthRef.current === 0) {
-      setCollapsed(true);
-    }
-    prevLengthRef.current = activities.length;
-  }, [isRunning, activities.length]);
-
-  if (activities.length === 0) return null;
-
-  // Force-expand when the active search query matches inside any of this
-  // section's activity summaries — otherwise marks would be silently
-  // hidden behind the collapsed header and the user would see a non-zero
-  // counter with no visible highlight. Match against the same relativized
-  // text we render, not the raw summary.
-  const queryHasMatch =
-    !!searchQuery &&
-    activities.some((a) => {
-      if (!a.summary) return false;
-      const text = relativizePath(a.summary, worktreePath);
-      return text.toLowerCase().includes(searchQuery.toLowerCase());
-    });
-  const isExpanded = !collapsed || queryHasMatch;
-
-  return (
-    <div className={styles.toolActivities} aria-live="polite" aria-atomic="true">
-      <div className={styles.turnSummary}>
-        <div
-          className={styles.turnHeader}
-          role="button"
-          tabIndex={0}
-          onClick={() => setCollapsed(!collapsed)}
-          onKeyDown={(e) => {
-            if (e.key === "Enter" || e.key === " ") {
-              e.preventDefault();
-              setCollapsed(!collapsed);
-            }
-          }}
-        >
-          <span className={styles.toolChevron}>
-            {isExpanded ? "⌄" : "›"}
-          </span>
-          <span className={styles.turnLabel}>
-            {activities.length} tool call{activities.length !== 1 ? "s" : ""}
-            {isRunning && <span className={styles.inProgressNote}> in progress</span>}
-          </span>
-        </div>
-        {isExpanded && (
-          <div className={styles.turnActivities}>
-            {activities.map((act: ToolActivity) => (
-              <div key={act.toolUseId} className={styles.toolActivity}>
-                <div className={styles.toolHeader}>
-                  <span className={styles.toolName} style={{ color: toolColor(act.toolName) }}>{act.toolName}</span>
-                  {act.summary && (
-                    <span className={styles.toolSummary}>
-                      <HighlightedPlainText text={relativizePath(act.summary, worktreePath)} query={searchQuery} />
-                    </span>
-                  )}
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-});
-
-/**
- * Shows a progress bar for the current in-progress turn, only when
- * task-related tools are among the current activities. Disappears when
- * the turn finalises (tasks move into CompletedTurn rendering).
- */
-const CurrentTurnTaskProgress = memo(function CurrentTurnTaskProgress({
-  sessionId,
-}: {
-  sessionId: string;
-}) {
-  const completedTurns = useAppStore(
-    (s) => s.completedTurns[sessionId] ?? EMPTY_COMPLETED_TURNS
-  );
-  const toolActivities = useAppStore(
-    (s) => s.toolActivities[sessionId] ?? EMPTY_ACTIVITIES
-  );
-
-  const result = useMemo(
-    () => deriveTasks(completedTurns, toolActivities),
-    [completedTurns, toolActivities]
-  );
-
-  // Only render when the current turn has task tools
-  if (!hasTaskActivity(toolActivities) || result.totalCount === 0) return null;
-
-  return (
-    <TaskProgressBar completedCount={result.completedCount} totalCount={result.totalCount} />
-  );
-});
-
-/** Extract the @-query based on cursor position in the textarea. */
-function extractMentionQuery(text: string, cursorPos: number): string | null {
-  const before = text.slice(0, cursorPos);
-  const atIndex = before.lastIndexOf("@");
-  if (atIndex === -1) return null;
-  // The @ must be at start of input or preceded by whitespace.
-  if (atIndex > 0 && !/\s/.test(before[atIndex - 1])) return null;
-  const query = before.slice(atIndex + 1);
-  // If query contains whitespace, the mention is "closed".
-  if (/\s/.test(query)) return null;
-  return query;
-}
-
-// Separate component for input area to prevent full ChatPanel re-renders on every keystroke
-/** Convert a File/Blob to a base64 string (without the data: prefix). */
-function fileToBase64(file: Blob): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => {
-      const result = reader.result as string;
-      const base64 = result.split(",")[1] ?? "";
-      resolve(base64);
-    };
-    reader.onerror = reject;
-    reader.readAsDataURL(file);
-  });
-}
-
-function ChatInputArea({
-  onSend,
-  onStop,
-  isRunning,
-  isRemote,
-  selectedWorkspaceId,
-  sessionId,
-  repoId,
-  projectPath,
-  historyRef,
-  historyIndexRef,
-  draftRef,
-  onAttachmentContextMenu,
-  onAttachmentClick,
-}: {
-  onSend: (
-    content: string,
-    mentionedFiles?: Set<string>,
-    attachments?: AttachmentInput[],
-  ) => Promise<void>;
-  onStop: () => void | Promise<void>;
-  isRunning: boolean;
-  isRemote: boolean;
-  selectedWorkspaceId: string;
-  sessionId: string;
-  repoId: string | undefined;
-  projectPath: string | undefined;
-  historyRef: React.MutableRefObject<Record<string, string[]>>;
-  historyIndexRef: React.MutableRefObject<number>;
-  draftRef: React.MutableRefObject<string>;
-  onAttachmentContextMenu?: (
-    e: React.MouseEvent,
-    attachment: DownloadableAttachment,
-  ) => void;
-  onAttachmentClick?: (
-    e: React.MouseEvent,
-    attachment: DownloadableAttachment,
-  ) => void;
-}) {
-  const [chatInput, setChatInput] = useState("");
-  const [cursorPos, setCursorPos] = useState(0);
-  const [inputScrollTop, setInputScrollTop] = useState(0);
-  const [slashPickerIndex, setSlashPickerIndex] = useState(0);
-  const [slashPickerDismissed, setSlashPickerDismissed] = useState(false);
-  const [slashCommands, setSlashCommandsLocal] = useState<SlashCommand[]>([]);
-  const setSlashCommandsStore = useAppStore((s) => s.setSlashCommands);
-  const setSlashCommands = useCallback(
-    (cmds: SlashCommand[]) => {
-      setSlashCommandsLocal(cmds);
-      setSlashCommandsStore(selectedWorkspaceId, cmds);
-    },
-    [selectedWorkspaceId, setSlashCommandsStore],
-  );
-  const [filePickerIndex, setFilePickerIndex] = useState(0);
-  const [filePickerDismissed, setFilePickerDismissed] = useState(false);
-  const [workspaceFiles, setWorkspaceFiles] = useState<FileEntry[]>([]);
-  const [filesLoaded, setFilesLoaded] = useState(false);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
-  const { t } = useTranslation("chat");
-  const filesCache = useRef<Record<string, FileEntry[]>>({});
-  const mentionedFilesRef = useRef<Set<string>>(new Set());
-  const [pendingAttachments, setPendingAttachments] = useState<PendingAttachment[]>([]);
-  const [dragActive, setDragActive] = useState(false);
-  const [attachMenuOpen, setAttachMenuOpen] = useState(false);
-  const [contextPopoverOpen, setContextPopoverOpen] = useState(false);
-  const pluginRefreshToken = useAppStore((s) => s.pluginRefreshToken);
-  const openSettings = useAppStore((s) => s.openSettings);
-
-  const insertTranscript = useCallback((transcript: string) => {
-    const ta = textareaRef.current;
-    const start = ta?.selectionStart ?? cursorPos;
-    const end = ta?.selectionEnd ?? cursorPos;
-    setChatInput((currentInput) => {
-      const next = insertTranscriptAtSelection(
-        currentInput,
-        transcript,
-        start,
-        end,
-      );
-      setCursorPos(next.cursor);
-      requestAnimationFrame(() => {
-        const current = textareaRef.current;
-        if (!current) return;
-        current.focus();
-        current.selectionStart = current.selectionEnd = next.cursor;
-      });
-      return next.text;
-    });
-  }, [cursorPos]);
-
-  const focusVoiceProvider = useAppStore((s) => s.focusVoiceProvider);
-  const voice = useVoiceInput(
-    insertTranscript,
-    (providerId) => {
-      focusVoiceProvider(providerId);
-      openSettings("plugins");
-    },
-  );
-  const voiceErrorOpensSettings = shouldOpenVoiceSettingsForError(
-    voice.activeProvider,
-  );
-
-  // Esc cancels an active recording regardless of where focus is. The
-  // textarea's onKeyDown also handles Esc when it has focus; clicking
-  // the mic moves focus to the button, where Esc would otherwise just
-  // defocus it instead of stopping the recording.
-  //
-  // While recording, Esc is treated as exclusively "cancel recording" —
-  // we capture it ahead of bubbling handlers and stop propagation so
-  // it doesn't also close an unrelated popover/modal that happens to
-  // be open. Without this, the same keypress could cancel recording
-  // *and* dismiss the surrounding UI, which feels jumpy.
-  useEffect(() => {
-    if (voice.state !== "recording") return;
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key !== "Escape") return;
-      e.preventDefault();
-      e.stopPropagation();
-      voice.cancel();
-    };
-    window.addEventListener("keydown", onKey, true);
-    return () => window.removeEventListener("keydown", onKey, true);
-  }, [voice.state, voice.cancel]);
-
-  const handleInsertPinnedCommand = useCallback((commandText: string) => {
-    setChatInput((prev) => commandText + (prev ? " " + prev : ""));
-    textareaRef.current?.focus();
-  }, []);
-
-  // Per-session draft storage: save input when switching away,
-  // restore when switching back.
-  const draftsRef = useRef<Record<string, string>>({});
-  const prevSessionRef = useRef(sessionId);
-  useEffect(() => {
-    const prev = prevSessionRef.current;
-    if (prev !== sessionId) {
-      // Save draft for the session we're leaving.
-      draftsRef.current[prev] = chatInput;
-      // Restore draft for the session we're entering.
-      setChatInput(draftsRef.current[sessionId] ?? "");
-      prevSessionRef.current = sessionId;
-      // Reset file picker and attachment state for new session.
-      setFilesLoaded(false);
-      setWorkspaceFiles([]);
-      mentionedFilesRef.current = new Set();
-      // Clear staged attachments so they don't leak across sessions.
-      setPendingAttachments((prev) => {
-        for (const a of prev) {
-          if (a.preview_url.startsWith("blob:")) URL.revokeObjectURL(a.preview_url);
-        }
-        return [];
-      });
-      voice.cancel();
-    }
-  }, [sessionId]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  // Auto-focus the textarea when switching or creating sessions.
-  useEffect(() => {
-    requestAnimationFrame(() => textareaRef.current?.focus());
-  }, [sessionId]);
-
-  // Consume prefill text (e.g. from rollback) and focus the textarea.
-  const chatInputPrefill = useAppStore((s) => s.chatInputPrefill);
-  const setChatInputPrefill = useAppStore((s) => s.setChatInputPrefill);
-  useEffect(() => {
-    if (chatInputPrefill) {
-      setChatInput(chatInputPrefill);
-      setChatInputPrefill(null);
-      // Focus and move cursor to end after React re-renders.
-      requestAnimationFrame(() => {
-        const ta = textareaRef.current;
-        if (ta) {
-          ta.focus();
-          ta.selectionStart = ta.selectionEnd = ta.value.length;
-        }
-      });
-    }
-  }, [chatInputPrefill, setChatInputPrefill]);
-
-  const refreshSlashCommands = useCallback(() => {
-    listSlashCommands(projectPath, selectedWorkspaceId)
-      .then(setSlashCommands)
-      .catch((e) => console.error("Failed to load slash commands:", e));
-  }, [pluginRefreshToken, projectPath, selectedWorkspaceId]);
-
-  useEffect(() => {
-    let cancelled = false;
-    listSlashCommands(projectPath, selectedWorkspaceId)
-      .then((cmds) => {
-        if (!cancelled) setSlashCommands(cmds);
-      })
-      .catch((e) => console.error("Failed to load slash commands:", e));
-    return () => {
-      cancelled = true;
-    };
-  }, [projectPath, selectedWorkspaceId]);
-
-  // Filter by the command-name token (text before the first whitespace) so the
-  // picker stays open while the user types arguments. This keeps the argument
-  // hint visible for native commands like `/plugin install …`.
-  const slashQuery = describeSlashQuery(chatInput);
-  const slashQueryToken = slashQuery?.token ?? null;
-  const slashHasArgs = slashQuery?.hasArgs ?? false;
-  const slashResults = useMemo(
-    () => (slashQueryToken === null ? [] : filterSlashCommands(slashCommands, slashQueryToken)),
-    [slashCommands, slashQueryToken],
-  );
-  const showSlashPicker = slashQueryToken !== null && slashResults.length > 0 && !slashPickerDismissed;
-
-  useEffect(() => {
-    setSlashPickerIndex(0);
-    setSlashPickerDismissed(false);
-  }, [slashQueryToken]);
-
-  // --- File mention picker ---
-
-  const loadFiles = useCallback(async () => {
-    if (filesCache.current[selectedWorkspaceId]) {
-      setWorkspaceFiles(filesCache.current[selectedWorkspaceId]);
-      setFilesLoaded(true);
-      return;
-    }
-    try {
-      const files = await listWorkspaceFiles(selectedWorkspaceId);
-      filesCache.current[selectedWorkspaceId] = files;
-      setWorkspaceFiles(files);
-      setFilesLoaded(true);
-    } catch (e) {
-      console.error("Failed to load workspace files:", e);
-    }
-  }, [selectedWorkspaceId]);
-
-  const mentionQuery = extractMentionQuery(chatInput, cursorPos);
-  const mentionResults = useMemo(
-    () => (mentionQuery === null ? [] : matchFiles(workspaceFiles, mentionQuery)),
-    [workspaceFiles, mentionQuery],
-  );
-  const showFilePicker =
-    mentionQuery !== null && mentionResults.length > 0 && !filePickerDismissed && filesLoaded;
-
-  // Lazy-load file list on first @ trigger.
-  useEffect(() => {
-    if (mentionQuery !== null && !filesLoaded) {
-      loadFiles();
-    }
-  }, [mentionQuery, filesLoaded, loadFiles]);
-
-  // Reset picker index when query changes.
-  useEffect(() => {
-    setFilePickerIndex(0);
-    setFilePickerDismissed(false);
-  }, [mentionQuery]);
-
-  const insertFileMention = useCallback(
-    (file: FileEntry) => {
-      const before = chatInput.slice(0, cursorPos);
-      const atIndex = before.lastIndexOf("@");
-      const after = chatInput.slice(cursorPos);
-      const mention = `@${file.path}`;
-      // Directories: no trailing space so the user can keep narrowing.
-      // Files: add a trailing space to close the mention.
-      const suffix = file.is_directory ? "" : " ";
-      const newText = before.slice(0, atIndex) + mention + suffix + after;
-      setChatInput(newText);
-      const newCursor = atIndex + mention.length + suffix.length;
-      setCursorPos(newCursor);
-      if (!file.is_directory) {
-        mentionedFilesRef.current.add(file.path);
-      }
-      requestAnimationFrame(() => {
-        const ta = textareaRef.current;
-        if (ta) {
-          ta.selectionStart = ta.selectionEnd = newCursor;
-          ta.focus();
-        }
-      });
-    },
-    [chatInput, cursorPos],
-  );
-
-  // Auto-resize textarea based on content
-  useEffect(() => {
-    const textarea = textareaRef.current;
-    if (!textarea) return;
-
-    // Reset height to auto to get the correct scrollHeight
-    textarea.style.height = "auto";
-    // Set height to scrollHeight; CSS max-height will cap it
-    textarea.style.height = `${textarea.scrollHeight}px`;
-  }, [chatInput]);
-
-  // -- Attachment helpers --
-
-  const addAttachment = useCallback(async (file: Blob, filename: string, textContent?: string) => {
-    if (isRemote) return; // Attachments not supported over remote transport
-    if (!SUPPORTED_ATTACHMENT_TYPES.has(file.type)) {
-      console.warn(`Unsupported file type: ${file.type}`);
-      return;
-    }
-    const isPdf = SUPPORTED_DOCUMENT_TYPES.has(file.type);
-    const isImage = SUPPORTED_IMAGE_TYPES.has(file.type);
-    const isText = isTextFile(file.type);
-    const sizeLimit = maxSizeFor(file.type);
-    if (file.size > sizeLimit) {
-      console.warn(
-        `File too large: ${(file.size / 1024 / 1024).toFixed(1)} MB (max ${(sizeLimit / 1024 / 1024).toFixed(1)} MB)`,
-      );
-      return;
-    }
-    const data_base64 = await fileToBase64(file);
-    let preview_url: string;
-    if (isPdf) {
-      const { generatePdfThumbnail } = await import("../../utils/pdfThumbnail");
-      preview_url = await generatePdfThumbnail(await file.arrayBuffer()).catch(() => "");
-      if (!preview_url) return;
-    } else if (isImage) {
-      preview_url = URL.createObjectURL(file);
-    } else {
-      preview_url = "";
-    }
-    const att: PendingAttachment = {
-      id: crypto.randomUUID(),
-      filename,
-      media_type: file.type,
-      data_base64,
-      preview_url,
-      size_bytes: file.size,
-      text_content: isText ? (textContent ?? await file.text()) : null,
-    };
-    setPendingAttachments((prev) => {
-      if (prev.length >= MAX_ATTACHMENTS) {
-        if (preview_url.startsWith("blob:")) URL.revokeObjectURL(preview_url);
-        return prev;
-      }
-      return [...prev, att];
-    });
-  }, [isRemote]);
-
-  const removeAttachment = useCallback((id: string) => {
-    setPendingAttachments((prev) => {
-      const att = prev.find((a) => a.id === id);
-      if (att?.preview_url.startsWith("blob:")) URL.revokeObjectURL(att.preview_url);
-      return prev.filter((a) => a.id !== id);
-    });
-  }, []);
-
-  // Track current attachments in a ref so the unmount cleanup always
-  // revokes the latest blob URLs (not the stale initial-render snapshot).
-  const pendingAttachmentsRef = useRef(pendingAttachments);
-  pendingAttachmentsRef.current = pendingAttachments;
-  useEffect(() => {
-    return () => {
-      for (const a of pendingAttachmentsRef.current) {
-        if (a.preview_url.startsWith("blob:")) URL.revokeObjectURL(a.preview_url);
-      }
-    };
-  }, []);
-
-  // Consume attachment prefill (e.g. from rollback) — convert the raw
-  // base64 data back into PendingAttachment objects with preview URLs.
-  const attachmentsPrefill = useAppStore((s) => s.pendingAttachmentsPrefill);
-  const setAttachmentsPrefill = useAppStore((s) => s.setPendingAttachmentsPrefill);
-  useEffect(() => {
-    if (!attachmentsPrefill || attachmentsPrefill.length === 0) return;
-    setAttachmentsPrefill(null);
-
-    (async () => {
-      for (const a of attachmentsPrefill) {
-        const bytes = base64ToBytes(a.data_base64);
-        const blob = new Blob([bytes], { type: a.media_type });
-        await addAttachment(blob, a.filename, a.text_content ?? undefined);
-      }
-    })().catch((e) => console.error("Failed to restore attachment prefill:", e));
-  }, [attachmentsPrefill, setAttachmentsPrefill, addAttachment]);
-
-  const handlePaste = useCallback(
-    (e: React.ClipboardEvent) => {
-      const items = e.clipboardData?.items;
-      if (!items) return;
-
-      for (const item of items) {
-        // Skip text/plain — pasting text should insert into the textarea,
-        // not create a file attachment.
-        if (item.type === "text/plain") continue;
-        // Some clipboard writers (notably `navigator.clipboard.write` with
-        // a ClipboardItem) expose the image both as a "string" item (its
-        // data URL) and a "file" item. We must check the file variant —
-        // getAsFile() returns null for string items, which would
-        // silently drop the paste.
-        if (item.kind !== "file") continue;
-        if (SUPPORTED_ATTACHMENT_TYPES.has(item.type)) {
-          e.preventDefault();
-          const file = item.getAsFile();
-          if (file) {
-            const defaultName = item.type === "application/pdf"
-              ? "pasted-document.pdf"
-              : "pasted-image.png";
-            addAttachment(file, file.name || defaultName);
-          }
-          return; // Only handle first attachment
-        }
-      }
-      // If no supported items, let the default text paste proceed.
-    },
-    [addAttachment],
-  );
-
-  // Tauri intercepts native file drops before they reach the webview's HTML5
-  // drag events. Use Tauri's onDragDropEvent to handle file drops, and fall
-  // through to readFileAsBase64 (which validates type + size on the Rust side).
-  //
-  // The handler references addAttachment via a ref to avoid re-registering the
-  // listener when the callback identity changes — re-registration causes a race
-  // where the old listener's async cleanup hasn't fired yet and the same drop
-  // event is processed by both the old and new listeners, duplicating files.
-  const addAttachmentRef = useRef(addAttachment);
-  addAttachmentRef.current = addAttachment;
-  const tauriDragListenerActive = useRef(false);
-
-  useEffect(() => {
-    if (isRemote) return;
-    let cancelled = false;
-    let unlisten: (() => void) | null = null;
-
-    import("@tauri-apps/api/webview").then(({ getCurrentWebview }) => {
-      if (cancelled) return;
-      return getCurrentWebview()
-        .onDragDropEvent((event) => {
-          if (cancelled) return;
-          if (event.payload.type === "enter" || event.payload.type === "over") {
-            setDragActive(true);
-          } else if (event.payload.type === "leave") {
-            setDragActive(false);
-          } else if (event.payload.type === "drop") {
-            setDragActive(false);
-            for (const filePath of event.payload.paths) {
-              readFileAsBase64(filePath)
-                .then((result) => {
-                  if (cancelled) return;
-                  const bytes = base64ToBytes(result.data_base64);
-                  const blob = new Blob([bytes], { type: result.media_type });
-                  addAttachmentRef.current(blob, result.filename, result.text_content ?? undefined);
-                })
-                .catch((err) =>
-                  console.warn("Skipped dropped file:", err),
-                );
-            }
-          }
-        })
-        .then((fn) => {
-          if (cancelled) {
-            fn();
-          } else {
-            unlisten = fn;
-            tauriDragListenerActive.current = true;
-          }
-        });
-    }).catch((err) => {
-      console.error(
-        "[drag-drop] Tauri native listener failed, falling back to HTML5:",
-        err,
-      );
-      tauriDragListenerActive.current = false;
-      setDragActive(false);
-    });
-
-    return () => {
-      cancelled = true;
-      tauriDragListenerActive.current = false;
-      unlisten?.();
-    };
-  }, [isRemote]);
-
-  // HTML5 file-drop fallback: activates only when the Tauri native handler
-  // failed to register (tauriDragListenerActive is false). The global dragover
-  // preventDefault is always active to suppress the browser's default
-  // file-navigation behavior.
-  useEffect(() => {
-    if (isRemote) return;
-
-    const preventNav = (e: DragEvent) => {
-      e.preventDefault();
-    };
-
-    const handleDragEnter = (e: DragEvent) => {
-      if (tauriDragListenerActive.current) return;
-      if (!e.dataTransfer?.types.includes("Files")) return;
-      e.preventDefault();
-      setDragActive(true);
-    };
-
-    const handleDragLeave = (e: DragEvent) => {
-      if (tauriDragListenerActive.current) return;
-      if (e.relatedTarget) return;
-      setDragActive(false);
-    };
-
-    const handleDrop = (e: DragEvent) => {
-      if (tauriDragListenerActive.current) return;
-      if (!e.dataTransfer?.types.includes("Files")) return;
-      e.preventDefault();
-      e.stopPropagation();
-      setDragActive(false);
-      const files = e.dataTransfer.files;
-      if (files.length === 0) return;
-      for (const file of Array.from(files)) {
-        void addAttachmentRef.current(file, file.name).catch((error) => {
-          console.error("[drag-drop] Failed to add dropped attachment:", error);
-        });
-      }
-    };
-
-    document.addEventListener("dragover", preventNav);
-    document.addEventListener("dragenter", handleDragEnter);
-    document.addEventListener("dragleave", handleDragLeave);
-    document.addEventListener("drop", handleDrop);
-
-    return () => {
-      document.removeEventListener("dragover", preventNav);
-      document.removeEventListener("dragenter", handleDragEnter);
-      document.removeEventListener("dragleave", handleDragLeave);
-      document.removeEventListener("drop", handleDrop);
-    };
-  }, [isRemote]);
-
-  const handleAttachClick = useCallback(async () => {
-    const selected = await open({ multiple: true });
-    if (!selected) return;
-    const paths = Array.isArray(selected) ? selected : [selected];
-    for (const filePath of paths) {
-      try {
-        const result = await readFileAsBase64(filePath);
-        const bytes = base64ToBytes(result.data_base64);
-        const blob = new Blob([bytes], { type: result.media_type });
-        await addAttachment(blob, result.filename, result.text_content ?? undefined);
-      } catch (err) {
-        console.error("Failed to read file:", err);
-      }
-    }
-  }, [addAttachment]);
-
-  const handleSend = () => {
-    voice.cancel();
-    // Only include files whose @path tokens are still in the text, so that
-    // removed references don't get expanded.
-    const activeFiles = new Set<string>();
-    for (const path of mentionedFilesRef.current) {
-      if (chatInput.includes(`@${path}`)) {
-        activeFiles.add(path);
-      }
-    }
-    const files = activeFiles.size > 0 ? activeFiles : undefined;
-    const attachmentPayload =
-      pendingAttachments.length > 0
-        ? pendingAttachments.map((a) => ({
-            filename: a.filename,
-            media_type: a.media_type,
-            data_base64: a.data_base64,
-            text_content: a.text_content ?? undefined,
-          }))
-        : undefined;
-    onSend(chatInput, files, attachmentPayload);
-    setChatInput("");
-    // Revoke blob URLs to free memory (data: URLs don't need cleanup).
-    for (const a of pendingAttachments) {
-      if (a.preview_url.startsWith("blob:")) URL.revokeObjectURL(a.preview_url);
-    }
-    setPendingAttachments([]);
-    mentionedFilesRef.current = new Set();
-  };
-
-  const planMode = useAppStore(
-    (s) => s.planMode[sessionId] ?? false,
-  );
-  const setPlanMode = useAppStore((s) => s.setPlanMode);
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === "Escape" && voice.state === "recording") {
-      e.preventDefault();
-      voice.cancel();
-      return;
-    }
-
-    // Shift+Tab: toggle plan mode
-    if (e.key === "Tab" && e.shiftKey) {
-      e.preventDefault();
-      setPlanMode(sessionId, !planMode);
-      return;
-    }
-
-    // File mention picker navigation (takes priority over slash picker)
-    if (showFilePicker) {
-      if (e.key === "ArrowDown") {
-        e.preventDefault();
-        setFilePickerIndex((i) => Math.min(i + 1, mentionResults.length - 1));
-        return;
-      }
-      if (e.key === "ArrowUp") {
-        e.preventDefault();
-        setFilePickerIndex((i) => Math.max(i - 1, 0));
-        return;
-      }
-      if (e.key === "Enter" && !e.shiftKey) {
-        e.preventDefault();
-        const result = mentionResults[filePickerIndex];
-        if (result) insertFileMention(result.file);
-        return;
-      }
-      if (e.key === "Tab" && !e.shiftKey) {
-        e.preventDefault();
-        const result = mentionResults[filePickerIndex];
-        if (result) insertFileMention(result.file);
-        return;
-      }
-      if (e.key === "Escape") {
-        e.preventDefault();
-        setFilePickerDismissed(true);
-        return;
-      }
-    }
-
-    // Slash command picker navigation
-    if (showSlashPicker) {
-      if (e.key === "ArrowDown") {
-        e.preventDefault();
-        setSlashPickerIndex((i) => Math.min(i + 1, slashResults.length - 1));
-        return;
-      }
-      if (e.key === "ArrowUp") {
-        e.preventDefault();
-        setSlashPickerIndex((i) => Math.max(i - 1, 0));
-        return;
-      }
-      if (e.key === "Enter" && !e.shiftKey) {
-        e.preventDefault();
-        const cmd = slashResults[slashPickerIndex];
-        if (cmd) {
-          // If the user has already typed arguments after the command name,
-          // keep what they typed; otherwise substitute the canonical name.
-          const send = slashHasArgs ? chatInput : "/" + cmd.name;
-          onSend(send);
-          setChatInput("");
-          // Native commands record their canonical name from inside the
-          // handleSend dispatcher; record here only for file-based commands
-          // that go straight to the agent.
-          if (!cmd.kind) {
-            recordSlashCommandUsage(selectedWorkspaceId, cmd.name)
-              .then(refreshSlashCommands)
-              .catch((e) => console.error("Failed to record slash command usage:", e));
-          }
-        }
-        return;
-      }
-      if (e.key === "Tab" && !e.shiftKey) {
-        e.preventDefault();
-        const cmd = slashResults[slashPickerIndex];
-        if (cmd) {
-          setChatInput("/" + cmd.name + " ");
-          setSlashPickerDismissed(true);
-        }
-        return;
-      }
-      if (e.key === "Escape") {
-        e.preventDefault();
-        setSlashPickerDismissed(true);
-        return;
-      }
-    }
-
-    if (e.key === "Enter" && !e.shiftKey) {
-      e.preventDefault();
-      handleSend();
-      return;
-    }
-
-    // History navigation with arrow keys
-    const history = historyRef.current[sessionId] ?? [];
-    if (history.length === 0) return;
-
-    if (e.key === "ArrowUp") {
-      e.preventDefault();
-      if (historyIndexRef.current === -1) {
-        draftRef.current = chatInput;
-        historyIndexRef.current = history.length - 1;
-      } else if (historyIndexRef.current > 0) {
-        historyIndexRef.current -= 1;
-      }
-      setChatInput(history[historyIndexRef.current]);
-    } else if (e.key === "ArrowDown") {
-      e.preventDefault();
-      if (historyIndexRef.current === -1) return;
-      if (historyIndexRef.current < history.length - 1) {
-        historyIndexRef.current += 1;
-        setChatInput(history[historyIndexRef.current]);
-      } else {
-        historyIndexRef.current = -1;
-        setChatInput(draftRef.current);
-      }
-    }
-  };
-
-  const showUltrathinkOverlay = hasUltrathink(chatInput);
-
-  return (
-    <div
-      className={`${styles.inputArea}${dragActive ? ` ${styles.inputDragActive}` : ""}`}
-    >
-      {showFilePicker && (
-        <FileMentionPicker
-          results={mentionResults}
-          selectedIndex={filePickerIndex}
-          onSelect={insertFileMention}
-          onHover={setFilePickerIndex}
-        />
-      )}
-      {showSlashPicker && (
-        <SlashCommandPicker
-          commands={slashResults}
-          selectedIndex={slashPickerIndex}
-          onSelect={(cmd) => {
-            const send = slashHasArgs ? chatInput : "/" + cmd.name;
-            onSend(send);
-            setChatInput("");
-            if (!cmd.kind) {
-              recordSlashCommandUsage(selectedWorkspaceId, cmd.name)
-                .then(refreshSlashCommands)
-                .catch((e) => console.error("Failed to record slash command usage:", e));
-            }
-          }}
-          onHover={setSlashPickerIndex}
-        />
-      )}
-      {pendingAttachments.length > 0 && (
-        <div className={styles.attachmentStrip}>
-          {pendingAttachments.map((att) => (
-            <div key={att.id} className={styles.attachmentThumb} title={att.filename}>
-              {isTextFile(att.media_type) ? (
-                <div className={styles.textFileBadge}>
-                  <FileText size={16} />
-                  <span className={styles.textFileName}>{att.filename}</span>
-                  <span className={styles.textFileSize}>
-                    {att.size_bytes < 1024
-                      ? `${att.size_bytes} B`
-                      : `${(att.size_bytes / 1024).toFixed(0)} KB`}
-                  </span>
-                </div>
-              ) : (
-                <img
-                  src={att.preview_url}
-                  alt={att.filename}
-                  onClick={(e) => {
-                    // PDFs also render as an <img> here (preview_url is a blob
-                    // URL of the first-page thumbnail), but their data_base64
-                    // is PDF bytes — not renderable inside an <img>. Only open
-                    // the lightbox for actual image MIME types.
-                    if (!att.media_type.startsWith("image/")) return;
-                    onAttachmentClick?.(e, {
-                      filename: att.filename,
-                      media_type: att.media_type,
-                      data_base64: att.data_base64,
-                    });
-                  }}
-                  onContextMenu={(e) =>
-                    onAttachmentContextMenu?.(e, {
-                      filename: att.filename,
-                      media_type: att.media_type,
-                      data_base64: att.data_base64,
-                    })
-                  }
-                />
-              )}
-              <button
-                className={styles.attachmentRemove}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  removeAttachment(att.id);
-                }}
-                title={t("remove_attachment")}
-              >
-                <X size={12} />
-              </button>
-            </div>
-          ))}
-        </div>
-      )}
-      <PinnedCommandsBar
-        repoId={repoId}
-        slashCommands={slashCommands}
-        onInsertCommand={handleInsertPinnedCommand}
-      />
-      <div className={styles.inputTextWrap}>
-        {showUltrathinkOverlay && (
-          <div className={styles.inputHighlight} aria-hidden="true">
-            <div style={{ transform: `translateY(-${inputScrollTop}px)` }}>
-              {renderUltrathinkText(chatInput, {
-                animated: true,
-                styles: {
-                  ultrathinkChar: styles.ultrathinkChar,
-                  ultrathinkCharAnimated: styles.ultrathinkCharAnimated,
-                },
-              })}
-            </div>
-          </div>
-        )}
-        <textarea
-          ref={textareaRef}
-          // data-chat-input is the stable selector used by the global focus
-          // shortcuts (Cmd+` and Cmd+0) in useKeyboardShortcuts.ts to move
-          // focus into the prompt from anywhere in the app.
-          data-chat-input
-          className={`${styles.input}${planMode ? ` ${styles.inputPlanMode}` : ""}${
-            showUltrathinkOverlay ? ` ${styles.inputWithHighlight}` : ""
-          }`}
-          value={chatInput}
-          onChange={(e) => {
-            setChatInput(e.target.value);
-            setCursorPos(e.target.selectionStart ?? 0);
-            setInputScrollTop(e.target.scrollTop);
-          }}
-          onSelect={(e) => {
-            setCursorPos((e.target as HTMLTextAreaElement).selectionStart ?? 0);
-          }}
-          onScroll={(e) => {
-            setInputScrollTop((e.target as HTMLTextAreaElement).scrollTop);
-          }}
-          onKeyDown={handleKeyDown}
-          onPaste={handlePaste}
-          placeholder={isRunning ? t("composer_placeholder_queued") : t("composer_placeholder_idle")}
-        />
-      </div>
-      <div className={styles.inputControls}>
-        <div className={styles.inputControlsLeft}>
-          <div className={styles.attachBtnWrap}>
-            <button
-              className={`${styles.attachBtn} ${attachMenuOpen ? styles.attachBtnActive : ""}`}
-              onClick={() => setAttachMenuOpen((v) => !v)}
-              title={t("add_files_connectors")}
-            >
-              <Plus size={16} />
-            </button>
-            {attachMenuOpen && (
-              <AttachMenu
-                repoId={repoId}
-                onAttachFiles={() => {
-                  setAttachMenuOpen(false);
-                  handleAttachClick();
-                }}
-                onClose={() => setAttachMenuOpen(false)}
-                isRemote={isRemote}
-              />
-            )}
-          </div>
-          <ComposerToolbar
-            sessionId={sessionId}
-            disabled={isRunning}
-          />
-        </div>
-        <div className={styles.inputControlsRight}>
-          <SegmentedMeter
-            sessionId={sessionId}
-            onClick={() => setContextPopoverOpen((v) => !v)}
-          />
-          {voice.state === "recording" && (
-            <div className={styles.voiceRecordingStatus} aria-live="polite">
-              <span className={styles.voiceWaveform} aria-hidden="true">
-                <span />
-                <span />
-                <span />
-              </span>
-              <span>{formatElapsedSeconds(voice.elapsedSeconds)}</span>
-            </div>
-          )}
-          {voice.state === "starting" && (
-            <div className={styles.voiceStatusText} aria-live="polite">
-              <LoaderCircle
-                size={12}
-                className={styles.voiceStatusSpinner}
-                aria-hidden="true"
-              />
-              <span>{t("voice_starting")}</span>
-            </div>
-          )}
-          {voice.state === "transcribing" && (
-            <div className={styles.voiceStatusText} aria-live="polite">
-              <LoaderCircle
-                size={12}
-                className={styles.voiceStatusSpinner}
-                aria-hidden="true"
-              />
-              <span>
-                {voice.activeProvider?.name
-                  ? t("voice_transcribing_with", { provider: voice.activeProvider.name })
-                  : t("voice_transcribing")}
-              </span>
-            </div>
-          )}
-          {voice.state === "error" && voice.error && (
-            voiceErrorOpensSettings ? (
-              <button
-                type="button"
-                className={styles.voiceErrorBtn}
-                onClick={() => openSettings("plugins")}
-                title={voice.error}
-              >
-                <AlertCircle size={12} className={styles.voiceErrorIcon} aria-hidden="true" />
-                <span className={styles.voiceErrorText}>{voice.error}</span>
-              </button>
-            ) : (
-              <button
-                type="button"
-                className={styles.voiceErrorBtn}
-                onClick={() => voice.cancel()}
-                title={`${voice.error ?? ""}\n\n${t("voice_error_dismiss_hint")}`}
-              >
-                <AlertCircle size={12} className={styles.voiceErrorIcon} aria-hidden="true" />
-                <span className={styles.voiceErrorText}>{voice.error}</span>
-              </button>
-            )
-          )}
-          <button
-            type="button"
-            className={`${styles.voiceBtn} ${voice.state === "recording" ? styles.voiceBtnRecording : ""} ${voice.state === "transcribing" || voice.state === "starting" ? styles.voiceBtnTranscribing : ""}`}
-            onClick={() => {
-              if (voice.state === "recording") voice.stop();
-              else if (
-                voice.state === "transcribing" ||
-                voice.state === "starting"
-              )
-                voice.cancel();
-              else void voice.start();
-            }}
-            disabled={isRunning}
-            title={
-              voice.state === "recording"
-                ? t("voice_stop")
-                : voice.state === "transcribing"
-                  ? t("voice_discard")
-                  : voice.state === "starting"
-                    ? t("voice_cancel")
-                    : t("voice_input")
-            }
-            aria-label={
-              voice.state === "recording"
-                ? t("voice_stop")
-                : voice.state === "transcribing"
-                  ? t("voice_discard")
-                  : voice.state === "starting"
-                    ? t("voice_cancel")
-                    : t("voice_input")
-            }
-          >
-            {voice.state === "transcribing" ? (
-              <X size={16} />
-            ) : voice.state === "starting" ? (
-              <LoaderCircle size={16} className={styles.voiceStatusSpinner} />
-            ) : (
-              <Mic size={16} />
-            )}
-          </button>
-          <button
-            className={`${styles.sendBtn} ${isRunning ? styles.sendBtnStop : ""}`}
-            onClick={isRunning ? onStop : handleSend}
-            disabled={!isRunning && !chatInput.trim() && pendingAttachments.length === 0}
-            title={isRunning ? t("stop_agent") : t("send_message")}
-            aria-label={isRunning ? t("stop_agent") : t("send_message")}
-          >
-            {isRunning ? <Square size={16} /> : <Send size={16} />}
-          </button>
-          {contextPopoverOpen && (
-            <ContextPopover
-              sessionId={sessionId}
-              onClose={() => setContextPopoverOpen(false)}
-              onCompact={() => { onSend("/compact"); }}
-              onClear={() => { onSend("/clear"); }}
-            />
-          )}
-        </div>
-      </div>
     </div>
   );
 }
