+++ src/ui/src/components/chat/ScrollContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+/** Context to pass sticky-scroll handler into streaming sub-components. */
+export const ScrollContext = createContext<{
+  handleContentChanged: () => void;
+}>({ handleContentChanged: () => {} });
